+++ docker_daemon/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG - docker_daemon
 
-1.1.0/Unreleased
+1.1.0/2017-04-24
 ==================
 ### Changes